Conflicts:
	libraries/osx/build-osx-libs.sh
	source/ps/SavedGame.cpp
	source/scriptinterface/ScriptInterface.cpp

+++ source/gui/GUIutil.h
@@ -1,4 +1,4 @@
-/* Copyright (C) 2009 Wildfire Games.
+/* Copyright (C) 2014 Wildfire Games.
  * This file is part of 0 A.D.
  *
  * 0 A.D. is free software: you can redistribute it and/or modify
@@ -37,7 +37,6 @@ GUI util
 //  Includes / Compiler directives
 //--------------------------------------------------------
 #include "GUIbase.h"
-#include "ps/Parser.h"
 // TODO Gee: New
 #include "ps/Overlay.h"
 #include "CGUI.h"
@@ -236,7 +235,7 @@ public:
 		return __ParseString<T>(Value, tOutput);
 	}
 
-	static bool ParseColor(const CStrW& Value, CColor &tOutput, float DefaultAlpha);
+	static bool ParseColor(const CStrW& Value, CColor &tOutput, int DefaultAlpha);
 
 private:
